pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "palu-labs"
version = "0.1.0"
description = "Palu Labs"
authors = ["Matheus Palu <my@email.com>"]
package-mode = false

[tool.poetry.dependencies]
flask = "^3.0.2"
gunicorn = "^21.2.0"
python = "^3.9"

[tool.poetry.dev-dependencies]
black = "^24.3.0"
pre-commit = "^3.7.0"
pylint = "^3.1.0"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"

